v 4.7
Android v 4.7
Major Feature:
Now an agent or admin will be able to send the file from mobile
application to the server.
Maximum file size will be 6 MB if it is more that that then it will not
be uploaded.
Major Bug:
Now while creating the ticket agent or admin will get only activated
user,deactivated users will not come into the suggestion.
If the API is disabled from web then agent/admin won't be able to log in
to the app.In case if he is already logged in to the app
then it will show the error message.
If the URL is going to change then the agent/admin will get the error
message that the url or the credentials have been changed.
Major UI changes:
Added loader icon in create ticket and reply.
Changed the user suggestion structure.Now it will display the name,email
and profile picture.
Added clear cache data button,but if there is any issue from server then
only the button will enable.